+++ changelog.md
@@ -1,4 +1,9 @@
+# v1.1.1
+- Added compatibility with real time icon, color, and glow changes
+- Fixed cursor not showing on pause menu
+
 # v1.1.0
 - Fixed cursor disappearing after returning from outside of GD
+
 # v1.0.0
 - Initial release